Update the development environment
Updated the development environment for the worker, making it easier to
start the development with the required pieces that are needed to ensure
a good execution throught the worker.

There is a podman-compose file located under
development/podman-compose.yml that configures the required services and
configurations in order for the worker/rhcd communicate and deal with
file uploading and downloading.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>