Add setup for sos report
SOS Report needs a special file in the /etc/sos.extras.d/ folder to be
able to collect the log files for the worker. Since it won't have a
special plugin in sosreport due to RHEL 7 being in maintenance mode, we
will be using this strategy to make sure that sosreport will be able to
collect our logs.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>