Refactor specfile for building the worker package
In this commit we are refactoring slightly how we build our worker, to
make it output dependency information in the binary that will help QE to
identify the versions we are using during the builds.

It also has some minor improvements to the Makefile, and with that, both
the specfile and our development setup uses the same entrypoints for
builds.